requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
numpy